+++ libraries/provider_git_client.rb
@@ -1,7 +1,0 @@
-class Chef
-  class Provider
-    class GitClient < Chef::Provider::LWRPBase
-      include GitCookbook::Helpers
-    end
-  end
-end