+++ metadata.rb
@@ -6,7 +6,9 @@ description       'Installs git and/or sets up a Git server daemon'
 version           '10.1.1'
 source_url        'https://github.com/sous-chefs/git'
 issues_url        'https://github.com/sous-chefs/git/issues'
-chef_version      '>= 14'
+chef_version      '>= 15.3'
+
+depends 'ark'
 
 supports 'amazon'
 supports 'centos'