+++ recipes/server.rb
@@ -18,5 +18,4 @@
 
 git_service 'default' do
   service_base_path node['git']['server']['base_path']
-  action :create
 end