+++ recipes/source.rb
@@ -18,11 +18,10 @@
 
 # drive version from node attributes
 git_client 'default' do
-  provider Chef::Provider::GitClient::Source
   source_checksum node['git']['checksum']
   source_prefix node['git']['prefix']
   source_url format(node['git']['url'], version: node['git']['version'])
   source_use_pcre node['git']['use_pcre']
   source_version node['git']['version']
-  action :install
+  action :install_from_source
 end